+++ src/main/kotlin/com/meetgom/backend/model/domain/event_sheet/EventSheetTimeSlot.kt
@@ -1,8 +1,9 @@
-package com.meetgom.backend.model.domain
+package com.meetgom.backend.model.domain.event_sheet
 
 import com.meetgom.backend.entity.event_sheet.EventSheetEntity
 import com.meetgom.backend.entity.event_sheet.EventSheetTimeSlotEntity
 import com.meetgom.backend.entity.event_sheet.EventSheetTimeSlotPrimaryKey
+import com.meetgom.backend.model.domain.common.TimeZone
 import com.meetgom.backend.model.http.response.EventSheetTimeSlotResponse
 import com.meetgom.backend.utils.TimeUtils
 import com.meetgom.backend.utils.extends.*
