+++ src/main/kotlin/com/meetgom/backend/utils/extends/EventSheetTimeSlotListExtends.kt
@@ -1,6 +1,6 @@
 package com.meetgom.backend.utils.extends
 
-import com.meetgom.backend.model.domain.EventSheetTimeSlot
+import com.meetgom.backend.model.domain.event_sheet.EventSheetTimeSlot
 import java.time.temporal.ChronoUnit
 
 fun List<EventSheetTimeSlot>.alignTimeSlots(): List<EventSheetTimeSlot> {
