+++ src/main/kotlin/com/meetgom/backend/utils/extends/LocalDateTimeExtends.kt
@@ -1,6 +1,6 @@
 package com.meetgom.backend.utils.extends
 
-import com.meetgom.backend.model.domain.TimeZone
+import com.meetgom.backend.model.domain.common.TimeZone
 import java.time.*
 
 fun LocalDateTime.atTimeZone(timeZone: TimeZone): ZonedDateTime {
